Move IPromptTemplate to namespace Microsoft.SemanticKernel.TemplateEngine
